[ABS-2533] Remove showInvisibles layer
The showInvisibles layer was created in the collision-based-kerning branch as a place to hold glyphs that included outline representations for control chars like ZWJ.
In the process of merging that branch to master, the showInvisible layer was not merged, but the glyphs in question still had outlines. This commit removes those
outlines. The glyphs affected are those for U+061C, U+200C - U+200F, 202A - 202E, 2066 - 2069, 206C, and 206D.  [nobuild]